refactor: rename newCheckout namespace to reactCheckout and some translation fixes
