+++ assets/javascripts/front/reactCheckout/payments/pix/index.js
