+++ src/Block/ReactCheckout/AbstractPaymentMethodBlock.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Woocommerce\Pagarme\Block\NewCheckout;
+namespace Woocommerce\Pagarme\Block\ReactCheckout;
 
 use Woocommerce\Pagarme\Core;
 use Woocommerce\Pagarme\Model\Payment\AbstractPayment;
