+++ src/Model/Checkout.php
@@ -109,7 +109,7 @@ class Checkout
     public function process(WC_Order $wc_order = null, string $type = CheckoutTypes::TRANSPARENT_VALUE)
     {
         if (
-            (!Utils::is_request_ajax() && !Utils::isNewCheckoutRequest())
+            (!Utils::is_request_ajax() && !Utils::isReactCheckoutRequest())
             || Utils::server('REQUEST_METHOD') !== 'POST'
         ) {
             exit(0);
